Disallow submitting form if banding is invalid
